fix: lock version of ECR module
The canaries failed to create an install with this sandbox, with the
error:

```sh
Could not retrieve the list of available versions for provider
hashicorp/aws:
locked provider registry.terraform.io/hashicorp/aws 5.35.0 does not
match
configured version constraint >= 4.0.0, >= 4.66.1, >= 5.30.0, >= 5.37.0;
must
use terraform init -upgrade to allow selection of new versions
```

We have AWS v5.35 in our lockfile, but the installed version of the ECR
module requires at least v5.37.

I think this snuck up on us because:

1. We require ">= 1.3.2" of the ECR module.
1. ECR at "1.3.2" requires ">= 4.0" of the AWS provider, which matches
   our required_providers block.
1. ECR at "2.1.0" requires ">= 5.37" of the AWS provider. This is the
   version we see in the error message when the install fails.
1. As long as the `.terraform` directory persists in any given env, the
   version of ECR is effectively locked. and things will keep working.
1. If the `.terraform` directory is blown away, or you run `terraform
   init -upgrade`, you will get a newer version of ECR. Then you'll get
   this error.

This PR upgrades the required version of AWS to v5.43, and locks the ECR
module version to v2.1.0, to prevent this from happening
again. We may want to consider locking module versions generally to
avoid this situation in the future.